Remove everything not directly related to debug overlay
