--- src/OpenSage.Game/DebugOverlay/IDebuggablePoint.cs
@@ -1,14 +1,0 @@
-﻿using OpenSage.Graphics.Cameras;
-using OpenSage.Mathematics;
-
-namespace OpenSage.DebugOverlay
-{
-    public interface IDebuggablePoint
-    {
-        ColorRgbaF DisplayColor { get; }
-
-        Rectangle GetBoundingRectangle(Camera camera);
-
-        bool Intersects(in BoundingFrustum frustum);
-    }
-}
