--- src/OpenSage.Game/Pathfinder/GridPoint.cs
@@ -1,12 +1,0 @@
-﻿using System.Numerics;
-using OpenSage.DebugOverlay;
-
-namespace OpenSage.Pathfinder
-{
-    public class GridPoint : DebuggablePoint
-    {
-        public GridPoint(Vector3 position) : base(position)
-        {
-        }   
-    }
-}
