+++ src/OpenSage.Game/Scene3D.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using OpenSage.Data.Map;
-using OpenSage.DebugOverlay;
 using OpenSage.Graphics.Cameras;
 using OpenSage.Graphics.ParticleSystems;
 using OpenSage.Graphics.Rendering;
@@ -10,7 +9,6 @@ using OpenSage.Graphics.Rendering.Shadows;
 using OpenSage.Gui;
 using OpenSage.Logic;
 using OpenSage.Logic.Object;
-using OpenSage.Mathematics;
 using OpenSage.Scripting;
 using OpenSage.Settings;
 
@@ -25,8 +23,6 @@ namespace OpenSage
         private CameraInputState _cameraInputState;
 
         private readonly SelectionMessageHandler _selectionMessageHandler;
-
-        private readonly ActionMessageHandler _actionMessageHandler;
 
         private readonly DebugMessageHandler _debugMessageHandler;
 
@@ -73,7 +69,6 @@ namespace OpenSage
         private List<Player> _players;
         public Player LocalPlayer { get; private set; }
         public DebugOverlay.DebugOverlay DebugOverlay { get; private set; }
-        public Pathfinder.Pathfinder Pathfinder { get; }
 
         internal IEnumerable<AttachedParticleSystem> GetAllAttachedParticleSystems()
         {
@@ -121,10 +116,6 @@ namespace OpenSage
             game.InputMessageBuffer.Handlers.Add(_selectionMessageHandler);
             AddDisposeAction(() => game.InputMessageBuffer.Handlers.Remove(_selectionMessageHandler));
 
-            _actionMessageHandler = new ActionMessageHandler(game.Action);
-            game.InputMessageBuffer.Handlers.Add(_actionMessageHandler);
-            AddDisposeAction(() => game.InputMessageBuffer.Handlers.Remove(_actionMessageHandler));
-
             _cameraInputMessageHandler = new CameraInputMessageHandler();
             game.InputMessageBuffer.Handlers.Add(_cameraInputMessageHandler);
             AddDisposeAction(() => game.InputMessageBuffer.Handlers.Remove(_cameraInputMessageHandler));
@@ -142,11 +133,6 @@ namespace OpenSage
             _teams = teams.ToList();
             // TODO: This is completely wrong.
             LocalPlayer = _players.FirstOrDefault();
-
-            Pathfinder = new Pathfinder.Pathfinder(mapFile);
-            AddDisposeAction(() => Pathfinder.Dispose());
-            Pathfinder.RemoveGridPointsForGameObjects(GameObjects);
-            DebugOverlay.AddGridPoints(Pathfinder.PassabilPoints);
         }
 
         public void SetPlayers(IEnumerable<Player> players, Player localPlayer)
